# fix concurrent map panic when querying and committing concurrently